BAU: Move config.ts to shared-identity folder
