+++ src/main/ipv-stub/render-ipv-authorize.ts
@@ -1,6 +1,6 @@
 import { JWTPayload, ProtectedHeaderParameters } from "jose";
 import { renderPage } from "../template.ts";
-import config, { getTrustmarkUri } from "./config/config.ts";
+import config, { getTrustmarkUri } from "../shared-identity/config/config.ts";
 
 export default function renderIPVAuthorize(
   decodedHeader: ProtectedHeaderParameters,
