+++ src/main/shared-identity/config/config.ts
